+++ CHANGELOG.md
@@ -4,12 +4,18 @@
 
 ## Unreleased
 
+### Added
+
+- passwords and passphrases are now stored in the much safer IntelliJ Credential Store.
+
 ### Changed
+
 - removed deprecated API usages
 
 ## 2.0.3 - 2023-03-01
 
 ### Added
+
 - support for experimental new UI
 
 ### Changed